update: hiiq big numnber issue
